chore: 試合ページのボタン類をコンポーネントに分割
